Make User object extend SoenEA classes
IUser and User extend IDomainObject<IDField> and DomainObject<IDField>, respectively.
IDField represents the type that we use for the id field, i.e. the primary key of an entity.
It's useful in that it lets us set any type we want. We usually set it to Long, as it is in this case.
Version attribute is of type long also. I had set it to int but I changed it.

UserProxy extends DomainObjectProxy, which acts as a proxy to domain objects (i.e. User).
It has an innerObject variable, storing an instance of the domain object type (i.e. User).
If the innerObject hasn't been initialized yet, it uses the UserMapper to set it to a User object.

UserMapper extends GenericOutputMapper, which provides insert, update, and delete methods to be implemented.

See #16

+++ src/dom/model/user/tdg/UserTDG.java
@@ -115,12 +115,12 @@ public class UserTDG {
 		return ps.executeQuery();
 	}
 	
-	public static int insert(long id, int version, String username, String password) throws SQLException {
+	public static int insert(long id, long version, String username, String password) throws SQLException {
 		Connection con = DbRegistry.getDbConnection();
 		
 		PreparedStatement ps = con.prepareStatement(INSERT);
 		ps.setLong(1, id);
-		ps.setInt(2, version);
+		ps.setLong(2, version);
 		ps.setString(3, username);
 		ps.setString(4, password);
 		
@@ -130,14 +130,14 @@ public class UserTDG {
 		return result;
 	}
 	
-	public static int update(String username, String password, long id, int version) throws SQLException {
+	public static int update(String username, String password, long id, long version) throws SQLException {
 		Connection con = DbRegistry.getDbConnection();
 		
 		PreparedStatement ps = con.prepareStatement(UPDATE);
 		ps.setString(1, username);
 		ps.setString(2, password);
 		ps.setLong(3, id);
-		ps.setInt(4, version);
+		ps.setLong(4, version);
 		
 		int result = ps.executeUpdate();
 		ps.close();
@@ -145,12 +145,12 @@ public class UserTDG {
 		return result;
 	}
 	
-	public static int delete(long id, int version) throws SQLException {
+	public static int delete(long id, long version) throws SQLException {
 		Connection con = DbRegistry.getDbConnection();
 		
 		PreparedStatement ps = con.prepareStatement(DELETE);
 		ps.setLong(1, id);
-		ps.setInt(2, version);
+		ps.setLong(2, version);
 		
 		int result = ps.executeUpdate();
 		ps.close();
